stocks_products/requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
